Fix discount search query bug involving displayName
The columns returned from the discount search query
in DiscountDAOImpl had duplicates of displayName and
some others. This meant that the displayName property
of stores and products and chains was incorrectly mapped
so a product had the same displayName as it's store.

This is a temporary hack to solve the issue. We are
attempting to get a demo video out tonight of the app
so a quality solution will have to come later.